Divided player file into three separate components, made an interface for props, implemented previously used import/export practice
