+++ src/modules/player/components/SKPlayer/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./SKIframe";
+export * from "./SKVideo";
+export * from "./SKImage";
