+++ tsconfig.json
@@ -4,6 +4,7 @@
     "paths": {
       "@models/*": ["./models/*"],
       "@playerComponents/*": ["./src/modules/player/components/*"],
+      "@SKplayerComponents/*": ["./src/modules/player/components/SKPlayer/components*"],
     },
     "target": "es5",
     "lib": ["dom", "dom.iterable", "esnext"],
